refactor: remove state from api
Tracking different states for the same chain would be messy
and i do not see what we would need it for right now.

The way state is handled internally right now implies that
claims and capabilities will be added and not removed.
They stay in the state after it has been commited as long
as the object remains in memory.

We lack a way to load the state from the stored chain.
At some point we might also want to seed a new block
with a state different from the one commited last.

But for now this simplifies the API quite a bit.