wip: add claims about all recipients and try to verify
This currently does not work. It never finds any value in the
process incoming part.

What it should do is add claims for all recipients including their
head imprints if we got them.
When receiving a message it should then try and verify the own head
imprint.

This is not the final logic. Its just meant as one step towards
finding claims in other peoples chains.