CCAccount: register_peer and use addr in add_claim
register_peer will store the peers cc info so we can share
claims with them later.
add_claim now expects a reader argument with the addr of
the reader

This way we have high level functions in the plugin hooks
and push the implementation details of how we store the
info about peer claim chains down to helper functions

In order to keep the old tests in test_plugin that work
without the autocrypt specific semantics we allow handing
in the public key in register_peer rather than looking
it up from the peers claim chain.

This also allows us to still run these tests based on
dicts as stores.