expose cc_account.chain rather than .head
We used CCAccount as a drop in replacement for its chain
in a number of places. This only worked because all these
functions relied on was .head being defined.

While this is fancy it may also be confusing and it relies
on implementation details of ClaimChain.

Instead we now actually hand over a chain when we need one.