first steps towards a sync to cchttpserver
Now the FileStore that can sync.
That is it can send it's entire data to a remote server
and recieve a single record from the server.
It would have a local path and a remote url as parameters.
Right now the remote url is still set to a dev default.

One store should be used for my own claims.
And a different store should be used for the claims of others.
We don't want to combine the two in the same local dir.
Otherwise we would start uploading other peoples claims.
If they use a different server
that would become an indirect sync between the servers.
I don't think we want that for now.

One thing that is interesting about what we are building here
is that the upload will basically be a brute force upload-it-all.
However the retrieval will always start from a specific key.
Then the content of that claim will be processed
and maybe we need to get some more nodes afterwards.

That's also what made me think about a tighter integration
between FileStore and Sync.
Because the claimchain is build on top of a store
which receives all the read calls.
We will have to delegate those to the sync then.

So the next step is to use the files as a cache
and retrieve the data from the remote server on reads
if it is missing.