filestore: recv missing data from remote
store[key] will now lookup the missing data on the remote server.

This way we can easily integrate it with claimchain.
For other peoples chains the store basically acts as a local cache.

Uploading the own claims still requires an explicit send.

Also using the [] API in the tests now and made file_set and file_get private.