wip: exchange claims via cchttpserver
This is in a funny state. It is fully working. claims are being pushed to the server
and also retrieved and verified.

However all of that is happening automagically when committing a claim
or verifying the incoming claims (which are all verified right away).

We also used to access the CC Account based on the dir specified in the url
part of the mails in the tests.
That was an ugly design to begin with and now it does not work anymore
because we use proper urls now.

However since the CCAccount is a plugin we cannot get from the account
to the CCAccount and therefore cannot introspect what is going on in the tests.

So what we need next is a way to trigger the verification from the outside.
We can leave in the verification that happens all the time or remove it.

Either way - once we have a trigger for sync and verfiication
we can write more meaningful tests again.